chat-server/pyproject.toml
[project]
name = "AnotherMe-ChatServer"
version = "0.1.0"
description = "AnotherMeのチャットサーバー"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi[standard]>=0.123.0",
    "pre-commit>=4.4.0",
    "pydantic-settings>=2.12.0",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "transformers>=4.57.3",
]

[dependency-groups]
dev = [
    "mypy==1.19.0",
    "ruff==0.14.8",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py313"
line-length = 88  # Blackのデフォルトと同じ

[tool.ruff.lint]
select = [
    "E",      # pycodestyleエラー
    "W",      # pycodestyle警告
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # 関数内の未使用引数
    "T201",   # print文を許可しない
]
ignore = [
    "B008",  # 引数のデフォルト値で関数呼び出しを行わない
    "W191",  # インデントにタブが含まれる
    "B904",  # from eなしで例外を発生させることを許可(HTTPException用)
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
